Added week2 exercise file.
