+++ Lectures/src/main/java/com/yearup/week2/CompilationProgram.java
@@ -1,4 +1,6 @@
-class Main {
+package com.yearup.week2;
+
+class CompilationProgram {
 public static void main(String args[]) { 
 int t; //This program wouldn't run as the value is not initialized. local variables need to be always initialized. 
 System.out.println(t);
