+++ Lectures/src/main/java/com/yearup/week2/TypeCastingExercise1.java
