+++ Lectures/src/main/java/com/yearup/week2/TypeCastingExercise2.java
