+++ Lectures/src/main/java/com/yearup/week2/VariableInit.java
